Shut down SSH master process when leaving tmt
The SSH master process and master socket may nto stay active in
sutiations when tmt is invoked multiple times and reusing the guest.
Each new invocation will try to start its own SSH master process, with
the same socket path, and this new SSH process will be stuck.

Plus, it's simply a resource leak: tmt quits and leaves a SSH process
running.

Patch adds a counterpart of `wake()` methods, `suspend()`, as in
"prepare for quitting, but you may be revisited later". It is a softer
version of what finish do - we want to release resources with runtime
lifespan, but not all of them.

There is an idea that tmt should split `finish` into two parts, running
`finish` phases and "cleanup" for taking care of resources. Once that
finalizes, the release of SSH master socket should moved into the
"cleanup" step, although even there it would need to remain a softer
version of the resource cleanup. In other words, the method will
probably stay with us, it just would be called from slightly different
part of tmt.

Fixes #3520.